refactor(cli): move skill manifest builder into cli package
