+++ packages/cli/package.json
@@ -11,7 +11,7 @@
     "skills"
   ],
   "scripts": {
-    "build": "bun ../../scripts/build-skill-manifest.ts && bun ../../scripts/sync-cli-skills.ts && bun build src/index.ts --outfile dist/wm.js --target bun",
+    "build": "bun ./scripts/build-skill-manifest.ts && bun ../../scripts/sync-cli-skills.ts && bun build src/index.ts --outfile dist/wm.js --target bun",
     "dev": "bun build src/index.ts --outfile dist/wm.js --target bun --watch",
     "install:bin": "bun run build && sudo ln -sf \"$(pwd)/dist/wm.js\" /usr/local/bin/wm && echo 'wm installed to /usr/local/bin (waymark available via package bin)'",
     "uninstall:bin": "sudo rm -f /usr/local/bin/wm && echo 'wm removed from /usr/local/bin'",
